add .xlsx cell and text styling
