+++ src/listing/onDraw/exportToXLSX/defineXLSX/bodyStyle.js
@@ -1,0 +1,24 @@
+export default {
+    font: {
+        sz: 10,
+        color: {
+            rgb: null, // set in defineXLSX
+        },
+    },
+    fill: {
+        fgColor: {
+            rgb: 'FFeeeeee',
+        },
+    },
+    alignment: {
+        wrapText: true,
+    },
+    border: {
+        bottom: {
+            style: 'thick',
+            color: {
+                rgb: null, // set in defineXLSX
+            },
+        },
+    },
+};
