+++ src/listing/onDraw/exportToXLSX/defineXLSX/headerStyle.js
@@ -1,0 +1,11 @@
+export default {
+    font: {
+        bold: true
+    },
+    fill: {
+        fgColor: {rgb: 'FFcccccc'}
+    },
+    alignment: {
+        wrapText: true
+    }
+};
