+++ src/listing/onDraw/exportToXLSX/defineXLSX/workBook.js
@@ -1,0 +1,4 @@
+export default function workBook() {
+    this.SheetNames = [];
+    this.Sheets = {};
+}
