+++ src/util/clone.js
@@ -1,0 +1,35 @@
+export default function clone(obj) {
+    let copy;
+
+    //boolean, number, string, null, undefined
+    if ('object' != typeof obj || null == obj)
+        return obj;
+
+    //date
+    if (obj instanceof Date) {
+        copy = new Date();
+        copy.setTime(obj.getTime());
+        return copy;
+    }
+
+    //array
+    if (obj instanceof Array) {
+        copy = [];
+        for (var i = 0, len = obj.length; i < len; i++) {
+            copy[i] = clone(obj[i]);
+        }
+        return copy;
+    }
+
+    //object
+    if (obj instanceof Object) {
+        copy = {};
+        for (var attr in obj) {
+            if (obj.hasOwnProperty(attr))
+                copy[attr] = clone(obj[attr]);
+        }
+        return copy;
+    }
+
+    throw new Error('Unable to copy [obj]! Its type is not supported.');
+}
